Add fastify as a peer dependency
This PR adds fastify as a peer dependency so that it works with Yarn v2 without having to override the package.
```Error: fastify-express tried to access fastify, but it isn't declared in its dependencies; this makes the require call ambiguous and unsound.```
Yarn v2 enforces packages to list all dependencies that are used by that package. In this case, fastify is provided by the user so it should be a peer dependency.